change the types of some c++ functions, streamline getting the geometric mean

+++ denovonear/cpp/simulate.h
@@ -1,9 +1,9 @@
 #ifndef SIMULATE_H
 #define SIMULATE_H
 
-std::vector<double> _get_distances(std::vector<int> sites);
-bool _has_zero(std::vector<double> distances);
-double _geomean(std::vector<double> distances);
+std::vector<int> _get_distances(std::vector<int> sites);
+bool _has_zero(std::vector<int> distances);
+double _geomean(std::vector<int> distances);
 bool _halt_permutation(double p_val, int iterations, double z=10.0, double alpha=0.01);
 std::vector<double> _simulate_distribution(Chooser choices,
     int iterations, int de_novo_count);
